+++ placeholder/src/main/kotlin/io/github/kioba/placeholder/model/UserId.kt
@@ -2,5 +2,5 @@ package io.github.kioba.placeholder.model
 
 @JvmInline
 public value class UserId(
-  val value: Int,
+  val value: Long,
 )
