+++ placeholder/src/main/kotlin/io/github/kioba/placeholder/post/NetworkPost.kt
@@ -8,5 +8,5 @@ data class NetworkPost(
   @Json(name = "body") val body: String,
   @Json(name = "id") val id: Int,
   @Json(name = "title") val title: String,
-  @Json(name = "userId") val userId: Int
+  @Json(name = "userId") val userId: Long,
 )
